修改平台分类id

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/goods/goods/GoodsPageListVo.java
@@ -19,7 +19,7 @@ public class GoodsPageListVo {
     private Byte goodsType;
     private String goodsImg;
     private String goodsSn;
-    private Short catId;
+    private Integer catId;
     private String catName;
     private String sortName;
     private Integer sortId;
